requirements.txt
asgiref==3.3.1
Django==2.2.17
django-cors-headers==3.6.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
